Revert "Fix rerender on key pressing in member page"
This reverts commit 7192158341adb29b3f60c3aed7c027cc7c69124b.
