+++ src/pages/members/MemberPage.tsx
@@ -1,7 +1,11 @@
 import { useTitle } from '@hedvig-ui/hooks/use-title'
 import { MemberTabs } from 'features/member'
 import { useGetMemberInfo } from 'features/member/tabs/member-tab/hooks/use-get-member-info'
-import React from 'react'
+import {
+  FocusItems,
+  useNavigation,
+} from 'features/navigation/hooks/use-navigation'
+import React, { useEffect } from 'react'
 import { RouteComponentProps } from 'react-router'
 import { Page } from 'pages/routes'
 
@@ -15,11 +19,27 @@ const MemberPage: Page<
 
   useTitle(member ? `${member?.firstName} ${member?.lastName}` : 'Loading...')
 
+  const { focus, setFocus } = useNavigation()
+
+  useEffect(() => {
+    if (!focus) {
+      setFocus(FocusItems.Member.name)
+    }
+  }, [focus])
+
   if (!member) {
     return null
   }
 
-  return <MemberTabs {...props} member={member} />
+  return (
+    <MemberTabs
+      {...props}
+      member={member}
+      navigationAvailable={focus === FocusItems.Member.name}
+      setFocus={(value: string | null) => setFocus(value)}
+      focus={focus}
+    />
+  )
 }
 
 export default MemberPage
